test_SUITE: Fetch dependencies using Maven
Thus, we do not depend on a local clone of rabbitmq-java-client and we
do not need to maintain the list of its dependencies anymore. This
should fix the testsuite now that rabbitmq-java-client depends on
org.slf4j.slf4j-api.

A better solution would be to convert the example project used by the
testsuite to a Maven project.